fix: revert 3.3.6's vtable case.
