+++ test/cases/05-VirtualTables/in/test_vctable_select_test_function.in
@@ -158,6 +158,3 @@ select mavg(int_tag, 100) from test_vtable_select.vtb_virtual_ctb_full limit 5;
 select statecount(int_tag, "GT", 20) from test_vtable_select.vtb_virtual_ctb_full limit 5;
 select stateduration(int_tag, "GT", 20, 1a) from test_vtable_select.vtb_virtual_ctb_full limit 5;
 select twa(int_tag) from test_vtable_select.vtb_virtual_ctb_full limit 5;
-
-select count(tbname) from test_vtable_select.vtb_virtual_ctb_full;
-select count(1) from test_vtable_select.vtb_virtual_ctb_full;
