+++ test/cases/05-VirtualTables/in/test_vstable_select_test_function.in
@@ -153,6 +153,4 @@ select derivative(int_tag, 5a, 1) from test_vtable_select.vtb_virtual_stb order 
 select irate(int_tag) from test_vtable_select.vtb_virtual_stb order by 1 limit 5;
 select statecount(int_tag, "GT", 20) from test_vtable_select.vtb_virtual_stb order by 1 limit 5;
 select stateduration(int_tag, "GT", 20, 1a) from test_vtable_select.vtb_virtual_stb order by 1 limit 5;
-
-select count(tbname) from test_vtable_select.vtb_virtual_stb;
-select count(1) from test_vtable_select.vtb_virtual_stb;
+select twa(int_tag) from test_vtable_select.vtb_virtual_stb order by 1 limit 5;
