+++ test/cases/05-VirtualTables/test_vtable_query_cross_db_stb.py
@@ -218,11 +218,10 @@ class TestVtableQueryCrossDbStb:
         tdCom.compare_testcase_result(self.sqlFile, self.ansFile, testCase)
 
     def test_select_virtual_super_table(self):
-        """Query: v-stable crossdb porject query
+        """test select virtual super table.
 
-        1. test vstable select super table cross db projection
-        2. test vstable select super table cross db projection filter
-        3. test vstable select super table cross db projection timerange filter
+        1 test vstable select super table cross db projection
+        2 test vstable select super table cross db projection filter
 
         Since: v3.3.6.0
 
@@ -233,10 +232,9 @@ class TestVtableQueryCrossDbStb:
         History:
             - 2025-3-15 Jing Sima Created
             - 2025-5-6 Huo Hong Migrated to new test framework
-            - 2025-10-23 Jing Sima Split function test into another case
 
         """
         self.run_normal_query("test_vstable_select_test_projection")
         self.run_normal_query("test_vstable_select_test_projection_filter")
-        self.run_normal_query("test_vstable_select_test_projection_timerange_filter")
+        #self.test_normal_query("test_vstable_select_test_function")
